Support ROS2 component for the pose estimator
